pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent", "src/llm"]

[project]
name = "assetopsbench-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "INSTRUCTIONS.md"
requires-python = ">=3.12"
dependencies = [
    "couchdb3>=2.0.2",
    "fastmcp>=2.14.5",
    "mcp[cli]>=1.26.0",
    "pendulum>=3.2.0",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "numpy>=1.24",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "litellm>=1.83.0",
    "python-dotenv>=1.0",
    "scipy>=1.10.0",
]

[project.scripts]
plan-execute = "agent.cli:main"
iot-mcp-server = "servers.iot.main:main"
utilities-mcp-server = "servers.utilities.main:main"
fmsr-mcp-server = "servers.fmsr.main:main"
tsfm-mcp-server = "servers.tsfm.main:main"
wo-mcp-server = "servers.wo.main:main"
vibration-mcp-server = "servers.vibration.main:main"
battery-mcp-server = "servers.battery.main:main"


[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "anyio>=4.9.0",
    "pytest-anyio>=0.0.0",
    "psutil>=7.2.2",
]
# Optional heavy ML deps for the TSFM server.
# granite-tsfm (tsfm_public) must be installed separately:
#   pip install git+https://github.com/ibm-granite/granite-tsfm
tsfm = [
    "torch>=2.0",
    "transformers>=4.40,<4.58",
    "huggingface-hub>=0.34,<1.0",
]
# Optional heavy deps for the battery server (acctouhou pretrained TF/Keras model).
# Acctouhou's .h5 weights were saved with Keras 2 (TF 2.2-era) and can't be deserialized
# by Keras 3. We install tf_keras (the legacy Keras 2 package) alongside TF 2.16+ and
# set TF_USE_LEGACY_KERAS=1 at import time. This is the supported migration path.
battery = [
    "tensorflow>=2.16",
    "tf_keras>=2.16",
    "scikit-learn>=1.3",
]

[tool.uv]
package = true

[tool.pytest.ini_options]
pythonpath = ["src"]
norecursedirs = ["src/tmp"]
filterwarnings = [
    "ignore:Core Pydantic V1 functionality:UserWarning",
]

